Removing `/pages/api`
Adding our own backend as a proxy for GitHub will rate-limit the whole service
at once. Whereas if we fetch on the client each user will have independent (ideally)
rate limits.
I also removed SWR because the data is not that important that we need to revalidate
and if we diable revalidation of SWR then there is no point of having SWR